requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filetype==1.2.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.9
langsmith==0.4.8
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.48.1
numpy==2.3.2
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
streamlit==1.47.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
yarl==1.20.1
zstandard==0.23.0
